feat(#81): constraint add, remove column function fix
