+++ packages/validator/src/lib/handlers/columns.ts
@@ -1,7 +1,7 @@
 import {
   ColumnDataTypeInvalidError,
   ColumnLengthRequiredError,
-  ColumnNameInvalidError,
+  ColumnInvalidError,
   ColumnNameInvalidFormatError,
   ColumnNameIsReservedKeywordError,
   ColumnNameNotUniqueError,
@@ -45,13 +45,6 @@ export interface ColumnHandlers {
     columnId: Column['id'],
     newPosition: Column['ordinalPosition']
   ) => Database;
-  changeColumnNullable: (
-    database: Database,
-    schemaId: Schema['id'],
-    tableId: Table['id'],
-    columnId: Column['id'],
-    nullable: boolean
-  ) => Database;
 }
 
 export const columnHandlers: ColumnHandlers = {
@@ -62,8 +55,8 @@ export const columnHandlers: ColumnHandlers = {
     const table = schema.tables.find((t) => t.id === tableId);
     if (!table) throw new TableNotExistError(tableId);
 
-    const result = COLUMN.shape.name.safeParse(column.name);
-    if (!result.success) throw new ColumnNameInvalidError(column.name);
+    const result = COLUMN.safeParse(column);
+    if (!result.success) throw new ColumnInvalidError(column);
 
     if (column.isAutoIncrement) {
       const autoIncrementColumn = table.columns.find((c) => c.isAutoIncrement);
@@ -179,34 +172,30 @@ export const columnHandlers: ColumnHandlers = {
     const column = table.columns.find((c) => c.id === columnId);
     if (!column) throw new ColumnNotExistError(columnId);
 
-    const columnNotUnique = table.columns.find((c) => c.name === newName);
+    const columnNotUnique = table.columns.find((c) => c.name === newName && c.id !== columnId);
     if (columnNotUnique) throw new ColumnNameNotUniqueError(newName);
 
     const result = COLUMN.shape.name.safeParse(newName);
-    if (!result.success) throw new ColumnNameInvalidError(newName);
+    if (!result.success) throw new ColumnInvalidError({ name: newName });
+
+    const changeTables: Table[] = schema.tables.map((t) =>
+      t.id === tableId
+        ? {
+            ...t,
+            isAffected: true,
+            columns: t.columns.map((c) => (c.id === columnId ? { ...c, name: newName, isAffected: true } : c)),
+          }
+        : t
+    );
+
+    const changeSchemas: Schema[] = database.schemas.map((s) =>
+      s.id === schemaId ? { ...s, isAffected: true, tables: changeTables } : s
+    );
 
     return {
       ...database,
       isAffected: true,
-      schemas: database.schemas.map((s) =>
-        s.id === schemaId
-          ? {
-              ...s,
-              isAffected: true,
-              tables: s.tables.map((t) =>
-                t.id === tableId
-                  ? {
-                      ...t,
-                      isAffected: true,
-                      columns: t.columns.map((c) =>
-                        c.id === columnId ? { ...c, name: newName, isAffected: true } : c
-                      ),
-                    }
-                  : t
-              ),
-            }
-          : s
-      ),
+      schemas: changeSchemas,
     };
   },
   changeColumnType: (database, schemaId, tableId, columnId, dataType, lengthScale) => {
@@ -221,36 +210,19 @@ export const columnHandlers: ColumnHandlers = {
     const column = table.columns.find((c) => c.id === columnId);
     if (!column) throw new ColumnNotExistError(columnId);
 
-    return {
-      ...database,
-      isAffected: true,
-      schemas: database.schemas.map((s) =>
-        s.id === schemaId
-          ? {
-              ...s,
-              isAffected: true,
-              tables: s.tables.map((t) =>
-                t.id === tableId
-                  ? {
-                      ...t,
-                      isAffected: true,
-                      columns: t.columns.map((c) =>
-                        c.id === columnId
-                          ? {
-                              ...c,
-                              isAffected: true,
-                              dataType,
-                              lengthScale: lengthScale || c.lengthScale,
-                            }
-                          : c
-                      ),
-                    }
-                  : t
-              ),
-            }
-          : s
-      ),
-    };
+    const changeColumns: Column[] = table.columns.map((c) =>
+      c.id === columnId ? { ...c, isAffected: true, dataType, lengthScale: lengthScale || c.lengthScale } : c
+    );
+
+    const changeTables: Table[] = schema.tables.map((t) =>
+      t.id === tableId ? { ...t, isAffected: true, columns: changeColumns } : t
+    );
+
+    const changeSchemas: Schema[] = database.schemas.map((s) =>
+      s.id === schemaId ? { ...s, isAffected: true, tables: changeTables } : s
+    );
+
+    return { ...database, isAffected: true, schemas: changeSchemas };
   },
   changeColumnPosition: (database, schemaId, tableId, columnId, newPosition) => {
     // NOTE: 변경에 의해서도 이전과 이후가 같은 상황에 대해선 고려가 필요없는지?
@@ -264,115 +236,18 @@ export const columnHandlers: ColumnHandlers = {
     const column = table.columns.find((c) => c.id === columnId);
     if (!column) throw new ColumnNotExistError(columnId);
 
-    return {
-      ...database,
-      isAffected: true,
-      schemas: database.schemas.map((s) =>
-        s.id === schemaId
-          ? {
-              ...s,
-              isAffected: true,
-              tables: s.tables.map((t) =>
-                t.id === tableId
-                  ? {
-                      ...t,
-                      isAffected: true,
-                      columns: t.columns.map((c) =>
-                        c.id === columnId
-                          ? {
-                              ...c,
-                              isAffected: true,
-                              updatedAt: new Date(),
-                              ordinalPosition: newPosition,
-                            }
-                          : c
-                      ),
-                    }
-                  : t
-              ),
-            }
-          : s
-      ),
-    };
-  },
-  changeColumnNullable: (database, schemaId, tableId, columnId, nullable) => {
-    // NOTE: 변경에 의해서도 이전과 이후가 같은 상황에 대해선 고려가 필요없는지?
-    // NOTE: 컬럼에서 nullable 변경이 있는게 맞는지? constraint에서 해당 부분을 처리하는게 옳아보이는데
+    const changeColumns: Column[] = table.columns.map((c) =>
+      c.id === columnId ? { ...c, isAffected: true, ordinalPosition: newPosition } : c
+    );
 
-    const schema = database.schemas.find((s) => s.id === schemaId);
-    if (!schema) throw new SchemaNotExistError(schemaId);
+    const changeTables: Table[] = schema.tables.map((t) =>
+      t.id === tableId ? { ...t, isAffected: true, columns: changeColumns } : t
+    );
 
-    const table = schema.tables.find((t) => t.id === tableId);
-    if (!table) throw new TableNotExistError(tableId);
+    const changeSchemas: Schema[] = database.schemas.map((s) =>
+      s.id === schemaId ? { ...s, isAffected: true, tables: changeTables } : s
+    );
 
-    const column = table.columns.find((c) => c.id === columnId);
-    if (!column) throw new ColumnNotExistError(columnId);
-
-    const currentNullable = helper.isColumnNullable(table, columnId);
-    if (currentNullable === nullable) return database;
-
-    const updatedColumn: Column = {
-      ...column,
-      isAffected: true,
-    };
-
-    let currentDatabase = columnHandlers.deleteColumn(structuredClone(database), schemaId, tableId, columnId);
-
-    currentDatabase = columnHandlers.createColumn(structuredClone(currentDatabase), schemaId, tableId, updatedColumn);
-
-    if (!nullable) {
-      const updatedSchema = currentDatabase.schemas.find((s) => s.id === schemaId)!;
-      const updatedTable = updatedSchema.tables.find((t) => t.id === tableId)!;
-
-      const hasNotNull = updatedTable.constraints.some(
-        (constraint) => constraint.kind === 'NOT_NULL' && constraint.columns.some((cc) => cc.columnId === columnId)
-      );
-
-      if (!hasNotNull) {
-        const newConstraintId = `constraint_${Date.now()}_${Math.random().toString(36).substr(2, 9)}`;
-        const newConstraint: Constraint = {
-          id: newConstraintId,
-          tableId,
-          name: `nn_${column.name}`,
-          kind: 'NOT_NULL' as const,
-          checkExpr: null,
-          defaultExpr: null,
-          columns: [
-            {
-              id: `cc_${Date.now()}_${Math.random().toString(36).substr(2, 9)}`,
-              constraintId: newConstraintId,
-              columnId,
-              seqNo: 1,
-              isAffected: true,
-            },
-          ],
-          isAffected: true,
-        };
-
-        currentDatabase = {
-          ...currentDatabase,
-          isAffected: true,
-          schemas: currentDatabase.schemas.map((s) =>
-            s.id === schemaId
-              ? {
-                  ...s,
-                  isAffected: true,
-                  tables: s.tables.map((t) =>
-                    t.id === tableId
-                      ? {
-                          ...t,
-                          isAffected: true,
-                          constraints: [...t.constraints, newConstraint],
-                        }
-                      : t
-                  ),
-                }
-              : s
-          ),
-        };
-      }
-    }
-
-    return currentDatabase;
+    return { ...database, isAffected: true, schemas: changeSchemas };
   },
 };
